Fix training via GUI with correct report file generation
Before it appears as nobody either tested the code or attempted to debug
it. It did not run and failed at obvious problems. Fixed that, so it
works again, although the reports via GUI will be lacking the fairness
information, as the GUI does not allow setting up anything like
demographic groups.